+++ src/components/Footer/Footer.css
@@ -1,10 +1,14 @@
 footer.main-footer {
   height: 35px;
   width: 100%;
+
   box-sizing: border-box;
   background-color: var(--main-color);
   color: var(--font-color);
+
   display: flex;
   align-items: center;
   justify-content: center;
+
+  text-align: center;
 }
